chore: refactor to remove index
