+++ src/pages/User/content/UserCreatedDocuments.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { Flex, Heading } from 'theme-ui'
 import UserCreatedDocumentsItem from './UserCreatedDocumentsItem'
-import type { UserCreatedDocs } from '.'
+import type { UserCreatedDocs } from '../types'
 
 interface IProps {
   docs: UserCreatedDocs | undefined
