+++ src/pages/User/types.ts
@@ -1,0 +1,4 @@
+export interface UserCreatedDocs {
+  howtos: any
+  research: any
+}
